refactor: change from old to es6
