+++ src/modules/__relative.js
@@ -1,6 +1,6 @@
 import path, { dirname } from 'path';
 import { fileURLToPath } from 'url';
 
-const __relative = (relativePath, pathname) => path.join(dirname(fileURLToPath(relativePath)), pathname);
+const __relative = (relativePath, pathname) => path.join(dirname(fileURLToPath(relativePath)), pathname || '');
 
 export default __relative;
