+++ src/modules/__relativeFilename.js
@@ -1,0 +1,7 @@
+
+
+import { fileURLToPath } from 'url';
+
+const __relativeFilename = (relativePath) => fileURLToPath(relativePath);
+
+export default __relativeFilename;
